+++ Modules/Modules.lua
@@ -2,7 +2,7 @@
 
 local AddonName, Addon = ...
 
-local DFL = DethsLibLoader("DethsFrameLib", "1.0")
+local DFL = DethsLibLoader("DethsFrameLib", "1.1")
 
 -- Libs
 Addon.Libs = {
@@ -10,7 +10,7 @@ Addon.Libs = {
   LDB = LibStub("LibDataBroker-1.1"),
   LDBIcon = LibStub("LibDBIcon-1.0"),
   DBL = DethsLibLoader("DethsBagLib", "1.0"),
-  DCL = DethsLibLoader("DethsColorLib", "1.0"),
+  DCL = DethsLibLoader("DethsColorLib", "1.1"),
   DFL = DFL,
   DTL = DethsLibLoader("DethsTooltipLib", "1.0")
 }
@@ -38,13 +38,18 @@ Addon.Frames = {
   ParentFrame = DFL.ParentFrame:Create(),
   TitleFrame = DFL.ChildFrame:Create(),
   DejunkChildFrame = DFL.ChildFrame:Create(),
+  ProfileChildFrame = DFL.ChildFrame:Create(),
   TransportChildFrame = DFL.ChildFrame:Create()
 }
 
 -- /UI/Objects/
 Addon.Objects = {
   CurrencyField = DFL:NewObjectTable(),
+  FauxButton = DFL:NewObjectTable(),
+  FauxFrame = DFL:NewObjectTable(),
   ListButton = DFL:NewObjectTable(),
   ListFrame = DFL:NewObjectTable(),
-  OptionsFrame = DFL:NewObjectTable()
+  OptionsFrame = DFL:NewObjectTable(),
+  ProfileButton = DFL:NewObjectTable(),
+  ProfileFrame = DFL:NewObjectTable()
 }