+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-### 8.0.7
-* Added "Ignore Readable" for both selling and destroying
+### 8.0.8
+* Added profile support
+* Moved the `Item Tooltip` & `Minimap Icon` options to the `General` options frame
+* Moved the `Auto Sell` & `Safe Mode` options to the `Sell` options frame
+* Updated locales
 * Updated libraries